refactor: domain 패키지 내부에 하위 패키지를 두도록 변경
